requirements.txt
streamlit
openai
pinecone-client
PyMuPDF
tiktoken